src/requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
gunicorn==21.0.1
Pillow==10.0.0
qrcode==7.4.2
python-barcode==0.15.1
celery==5.3.1
redis==4.6.0
django-redis==5.3.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
